Implement BL-002 author enrichment flow
