+++ src/shared/state.js
@@ -1,4 +1,9 @@
-import { RUN_STATES, STATUS_TEXT, TARGET_COUNT_DEFAULT } from "./constants.js";
+import {
+  ENRICHMENT_STATES,
+  RUN_STATES,
+  STATUS_TEXT,
+  TARGET_COUNT_DEFAULT,
+} from "./constants.js";
 import {
   clampTargetCount,
   getFeedState,
@@ -9,15 +14,31 @@ import {
 
 const tabStates = new Map();
 
+function createEmptyEnrichmentState() {
+  return {
+    status: ENRICHMENT_STATES.idle,
+    totalPosts: 0,
+    processedPosts: 0,
+    totalAuthors: 0,
+    processedAuthors: 0,
+    currentAuthor: null,
+    currentPostIndex: 0,
+    lastMessage: null,
+    readyForDownload: false,
+  };
+}
+
 function createEmptyState() {
   return {
     itemsByFingerprint: new Map(),
+    enrichedItems: [],
     status: STATUS_TEXT.idle,
     runState: RUN_STATES.idle,
     targetCount: TARGET_COUNT_DEFAULT,
     noProgressCycles: 0,
     stalledWaitCount: 0,
     lastReason: null,
+    enrichment: createEmptyEnrichmentState(),
   };
 }
 
@@ -57,6 +78,8 @@ export function getSerializableState(tabId) {
     noProgressCycles: tabState.noProgressCycles,
     stalledWaitCount: tabState.stalledWaitCount,
     lastReason: tabState.lastReason,
+    enrichedItems: tabState.enrichedItems.map(({ fingerprint, ...item }) => item),
+    enrichment: { ...tabState.enrichment },
   };
 }
 
@@ -90,6 +113,7 @@ export function markFeedReady(tabId, feedFound) {
 export function startCrawler(tabId, targetCount) {
   const tabState = getOrCreateTabState(tabId);
   Object.assign(tabState, getStartState(targetCount));
+  resetEnrichmentState(tabState);
   return persistState(tabId);
 }
 
@@ -156,10 +180,93 @@ export function mergeNewItems(tabId, items) {
     addedCount += 1;
   }
 
+  if (addedCount > 0) {
+    resetEnrichmentState(tabState);
+  }
+
   return {
     addedCount,
     state: getSerializableState(tabId),
   };
+}
+
+export function startEnrichment(
+  tabId,
+  { totalPosts = 0, totalAuthors = 0, lastMessage = null } = {},
+) {
+  const tabState = getOrCreateTabState(tabId);
+  tabState.enrichedItems = [];
+  tabState.enrichment = {
+    status: ENRICHMENT_STATES.running,
+    totalPosts,
+    processedPosts: 0,
+    totalAuthors,
+    processedAuthors: 0,
+    currentAuthor: null,
+    currentPostIndex: 0,
+    lastMessage,
+    readyForDownload: false,
+  };
+  return persistState(tabId);
+}
+
+export function updateEnrichmentProgress(tabId, patch = {}) {
+  const tabState = getOrCreateTabState(tabId);
+  tabState.enrichment = {
+    ...tabState.enrichment,
+    ...patch,
+    status: ENRICHMENT_STATES.running,
+    readyForDownload: false,
+  };
+  return persistState(tabId);
+}
+
+export function completeEnrichment(tabId, enrichedItems, lastMessage) {
+  const tabState = getOrCreateTabState(tabId);
+  tabState.enrichedItems = Array.isArray(enrichedItems)
+    ? enrichedItems.map((item) => ({ ...item }))
+    : [];
+  tabState.enrichment = {
+    ...tabState.enrichment,
+    status: ENRICHMENT_STATES.completed,
+    processedPosts: tabState.enrichment.totalPosts,
+    processedAuthors: tabState.enrichment.totalAuthors,
+    currentAuthor: null,
+    currentPostIndex: tabState.enrichment.totalPosts,
+    lastMessage: lastMessage || "Enrichment completed.",
+    readyForDownload: true,
+  };
+  return persistState(tabId);
+}
+
+export function failEnrichment(tabId, lastMessage) {
+  const tabState = getOrCreateTabState(tabId);
+  tabState.enrichedItems = [];
+  tabState.enrichment = {
+    ...tabState.enrichment,
+    status: ENRICHMENT_STATES.failed,
+    currentAuthor: null,
+    lastMessage: lastMessage || "Enrichment failed.",
+    readyForDownload: false,
+  };
+  return persistState(tabId);
+}
+
+export function cancelEnrichment(tabId, lastMessage) {
+  const tabState = getOrCreateTabState(tabId);
+  tabState.enrichedItems = [];
+  tabState.enrichment = {
+    ...tabState.enrichment,
+    status: ENRICHMENT_STATES.cancelled,
+    currentAuthor: null,
+    lastMessage: lastMessage || "Enrichment cancelled.",
+    readyForDownload: false,
+  };
+  return persistState(tabId);
+}
+
+export function getEnrichedItems(tabId) {
+  return getOrCreateTabState(tabId).enrichedItems.map((item) => ({ ...item }));
 }
 
 export async function persistState(tabId) {
@@ -187,12 +294,20 @@ export async function hydrateStateFromStorage(tabId) {
     tabState.itemsByFingerprint.set(fingerprint, { ...item, fingerprint });
   }
 
+  tabState.enrichedItems = Array.isArray(storedState?.enrichedItems)
+    ? storedState.enrichedItems.map((item) => ({ ...item }))
+    : [];
+
   tabState.status = storedState?.status || STATUS_TEXT.idle;
   tabState.runState = storedState?.runState || RUN_STATES.idle;
   tabState.targetCount = clampTargetCount(storedState?.targetCount);
   tabState.noProgressCycles = storedState?.noProgressCycles || 0;
   tabState.stalledWaitCount = storedState?.stalledWaitCount || 0;
   tabState.lastReason = storedState?.lastReason || null;
+  tabState.enrichment = {
+    ...createEmptyEnrichmentState(),
+    ...(storedState?.enrichment || {}),
+  };
   tabStates.set(tabId, tabState);
 
   return getSerializableState(tabId);
@@ -209,4 +324,9 @@ export async function ensureHydratedState(tabId) {
 export async function clearTabState(tabId) {
   tabStates.delete(tabId);
   await chrome.storage.session.remove(buildTabStorageKey(tabId));
+}
+
+function resetEnrichmentState(tabState) {
+  tabState.enrichedItems = [];
+  tabState.enrichment = createEmptyEnrichmentState();
 }
