+++ test/panel.smoke.test.js
@@ -45,9 +45,11 @@ describe("floating panel helpers", () => {
     const markup = createPanelMarkup();
 
     expect(markup).toContain("harvester-header");
-    expect(markup).toContain("harvester-export");
+    expect(markup).toContain("harvester-export-raw");
+    expect(markup).toContain("harvester-export-enriched");
     expect(markup).toContain("harvester-start");
     expect(markup).toContain("harvester-stop");
+    expect(markup).toContain("harvester-enrichment-status");
     expect(markup).toContain("harvester-status-badge");
     expect(markup).toContain("harvester-activity-log");
     expect(markup).toContain("data-target-preset");
