+++ vite.config.js
@@ -11,6 +11,10 @@ export default defineConfig({
       input: {
         popup: resolve(__dirname, "popup.html"),
         background: resolve(__dirname, "src/background/background.js"),
+        "linkedin-content": resolve(
+          __dirname,
+          "src/content/linkedin/content.js",
+        ),
       },
       output: {
         entryFileNames: "assets/[name].js",
